Wayland: Weston: Add GetNativeSurface() hook
This allows wpebackend-rdk to obtain the native parent surface.

This api is used by wpebackend-rdk which in turn has a new api
that is called from GStreamer backend in webkit. It allows the
video surface created by waylandsink to be a subsurface of the
browser surface created by Weston compositorclient code.

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>