compositorclient: add SurfaceByName() api
SurfaceByName() is used on the IDisplay class to get the ISurface.

This api is used with Wayland Weston compositor client when
using waylandsink as the video sink in wpewebkit. I've moved the
implementation to the Implementation.h header which is shared between
Westeros and Weston.

wpebackend-rdk calls these apis to obtain the surface handle created by
Thunder, which is then passed to waylandsink as an external surface so
that the video surface is correctly created as a sub-surface
of the browser.

Also RPI compositorclient is updated to have a stub implementation.

v2: reworked to better map to IDisplay & ISUrface classes
v3: removed NativeSurface() and extract wl_surface from EGLNativeWindow

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>