displayinfo: Decouple callbacks invoking from Updated() execution context
The DisplayInfo::Updated() method, executed in the context of a
DisplayInfo plugin COM-RPC call, is responsible for invoking the
callbacks that clients (e.g. Netflix, DisneyPlus) have registered to be
notified about changes in the DisplayInfo settings (HDMI/HDCP state,
etc.).

The clients' callbacks will usually make DisplayInfo calls themselves,
since they will be refreshing their stated based on the fact that they
were notified. These DisplayInfo calls (which will go to the DisplayInfo
plugin) are being done again in the same execution context of this
component's DisplayInfo::Updated() method, which in turn is serving and
blocking a COM-RPC call from the DisplayInfo plugin (where resources,
locks, etc. are being held), creating a potential for deadlocks, which
then result in COM-RPC timeouts and inconsistencies.

In particular, these kind of issues are being observed in the
interaction between the Netflix plugin and the DisplayInfo plugin,
through this 'displayinfo' client library. It can be seen that the
Netflix code handling the DisplayInfo callbacks is only completing its
execution around 10 seconds after it started. This matches precisely the
COM-RPC timeout, and it's happening because the Netflix callbacks are
doing calls to the DisplayInfo plugin (through this 'displayinfo' client
library) which needs to grab some resources (e.g.  mutex) that are
currently being held by the DisplayInfo plugin code that triggered the
notifications and is still waiting for them to return, causing a
deadlock which is only unblocked when the original DisplayInfo plugin
COM-RPC call for doing the notification times out.

To fix this deadlock potential, make sure that the Updated()
notification from the DisplayInfo plugin returns ASAP, by decoupling the
callbacks invoking for the registered clients from the Updated() context
into a worker pool job. For achieving this, use the 'Core::DecoupledJob'
interface, which simply abstracts the launching of arbitrary jobs
(materialized as lambdas) served by the worker pool.

Signed-off-by: Ricardo Silva <ricardo.josilva@parceiros.nos.pt>